Move compute_block_sparsity to flash_attn_interface.py
